refactor: Further cleaning `utils` directory
