+++ src/components/Saturation.tsx
@@ -4,7 +4,7 @@ import Interactive from "./Interactive";
 
 import styles from "../styles.css";
 import { HSV } from "../types";
-import { hsvToHslString } from "../utils/conversions";
+import { hsvToHslString } from "../utils/convert";
 import formatClassName from "../utils/formatClassName";
 
 interface Props {
