+++ src/packages/hex.tsx
@@ -3,8 +3,8 @@ import React from "react";
 import ColorPicker from "../components/ColorPicker";
 import withColorModel from "../hocs/withColorModel";
 import { ColorModel, ColorPickerBaseProps } from "../types";
-import { hexToHsv, hsvToHex } from "../utils/conversions";
-import equal from "../utils/equalHex";
+import { equalHex } from "../utils/compare";
+import { hexToHsv, hsvToHex } from "../utils/convert";
 
 interface Props extends ColorPickerBaseProps {
   color: string;
@@ -15,7 +15,7 @@ const colorModel: ColorModel<string> = {
   defaultColor: { h: 0, s: 0, l: 0 },
   toHsv: hexToHsv,
   fromHsv: hsvToHex,
-  equal,
+  equal: equalHex,
 };
 
 const HexColorPicker: React.FC<Props> = withColorModel(ColorPicker, colorModel);
