+++ src/packages/hsv.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import ColorPicker from "../components/ColorPicker";
 import withColorModel from "../hocs/withColorModel";
 import { ColorModel, ColorPickerBaseProps, HSV } from "../types";
-import equal from "../utils/equalColorObjects";
+import { equalColorObjects } from "../utils/compare";
 
 interface Props extends ColorPickerBaseProps {
   color: string;
@@ -14,7 +14,7 @@ const colorModel: ColorModel<HSV> = {
   defaultColor: { h: 0, s: 0, v: 0 },
   toHsv: (hsv) => hsv,
   fromHsv: (hsv) => hsv,
-  equal,
+  equal: equalColorObjects,
 };
 
 const HsvColorPicker: React.FC<Props> = withColorModel(ColorPicker, colorModel);
