+++ src/packages/rgb.tsx
@@ -3,8 +3,8 @@ import React from "react";
 import ColorPicker from "../components/ColorPicker";
 import withColorModel from "../hocs/withColorModel";
 import { ColorModel, ColorPickerBaseProps, RGB } from "../types";
-import { rgbToHsv, hsvToRgb } from "../utils/conversions";
-import equal from "../utils/equalColorObjects";
+import { equalColorObjects } from "../utils/compare";
+import { rgbToHsv, hsvToRgb } from "../utils/convert";
 
 interface Props extends ColorPickerBaseProps {
   color: RGB;
@@ -15,7 +15,7 @@ const colorModel: ColorModel<RGB> = {
   defaultColor: { r: 0, g: 0, b: 0 },
   toHsv: rgbToHsv,
   fromHsv: hsvToRgb,
-  equal,
+  equal: equalColorObjects,
 };
 
 const RgbColorPicker: React.FC<Props> = withColorModel(ColorPicker, colorModel);
