+++ src/packages/rgbString.tsx
@@ -3,8 +3,8 @@ import React from "react";
 import ColorPicker from "../components/ColorPicker";
 import withColorModel from "../hocs/withColorModel";
 import { ColorModel, ColorPickerBaseProps } from "../types";
-import { rgbStringToHsv, hsvToRgbString } from "../utils/conversions";
-import equal from "../utils/equalColorObjects";
+import { equalColorObjects } from "../utils/compare";
+import { rgbStringToHsv, hsvToRgbString } from "../utils/convert";
 
 interface Props extends ColorPickerBaseProps {
   color: string;
@@ -15,7 +15,7 @@ const colorModel: ColorModel<string> = {
   defaultColor: "rgb(0, 0, 0)",
   toHsv: rgbStringToHsv,
   fromHsv: hsvToRgbString,
-  equal,
+  equal: equalColorObjects,
 };
 
 const RgbStringColorPicker: React.FC<Props> = withColorModel(ColorPicker, colorModel);
